Removidos métodos e handlers redundantes relacionados às queries de `Course`, `Module` e `Lesson`. Adicionados novos métodos e repositórios com otimização via Dapper e inclusão de consultas específicas. Atualizados responses e ajustes gerais para consistência das respostas.

+++ src/02-Application/Application/Features/Learning/Commands/DeleteEnrollment/DeleteEnrollmentCommandHandler.cs
@@ -10,6 +10,10 @@ public class DeleteEnrollmentCommandHandler(IEnrollmentRepository enrollmentRepo
 {
     public Task<Result> Handle(DeleteEnrollmentCommand command, CancellationToken cancellationToken)
     {
+        var result = enrollmentRepository.GetById(command.UserId);
+        if (result == null)
+            return Task.FromResult(Result.Failure("Enrollment not found", "Not Found"));       
+        
         enrollmentRepository.Delete(command.UserId);
         unitOfWork.CommitAsync();
         return Task.FromResult(Result.Success());
